zsign.cpp: Parse new command line options related to ad-hoc signing
Parse and act on the new command line options detailed below.
```
-a, --adhoc		Perform ad-hoc signature only.
-s, --single_inplace	Re-sign a single Mach-O binary in place. (incompatible with `-o`)
-2, --sha256_only	Serialize a single code directory that uses SHA256.
```

`--adhoc` enables the use of ad-hoc signature instead.  `--single-inplace`
allows to re-sign a single Mach-O executable file.

`--sha256_only` forces the serialization of a single SHA256-based code
directory, which becomes the principal code directory.  If this option is
not specified, the default is to have a (principal) SHA1 code directory
+ an (alternate) SHA256 one.